+++ src/pub/index.php
@@ -55,7 +55,7 @@ if (isset($_GET['json'])) {
 if (isset($_GET['json_tower'])) {
     $sensor = $_GET['sensor'];
     require(APP_BASE_PATH . '/fcn/weather/getCurrentTowerData.php');
-    $getData = new getCurrentTowerData(sprintf('%08d', $_GET['sensor']));
+    $getData = new getCurrentTowerData(sprintf('%08d', $sensor));
     echo json_encode($getData->getConditions());
     die();
 }